// gateway/src/proto/order.ts
export interface BeerReady {
  beer_id: number;
  beer_name: string;
  quantity: number;
}

export interface OrderComplete {
  order_id: string;
  beers: BeerReady[];
  total_beers: number;
}

export interface OrderEvent {
  beer_ready?: BeerReady | undefined;
  order_complete?: OrderComplete | undefined;
}
// ...
export const BeerReady: MessageFns<BeerReady> = {
  encode(message: BeerReady, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.beer_id !== 0) {
      writer.uint32(8).int32(message.beer_id);
    }
    if (message.beer_name !== "") {
      writer.uint32(18).string(message.beer_name);
    }
    if (message.quantity !== 0) {
      writer.uint32(24).int32(message.quantity);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BeerReady {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBeerReady();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.beer_id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.beer_name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.quantity = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BeerReady {
    return {
      beer_id: isSet(object.beerId)
        ? globalThis.Number(object.beerId)
        : isSet(object.beer_id)
        ? globalThis.Number(object.beer_id)
        : 0,
      beer_name: isSet(object.beerName)
        ? globalThis.String(object.beerName)
        : isSet(object.beer_name)
        ? globalThis.String(object.beer_name)
        : "",
      quantity: isSet(object.quantity) ? globalThis.Number(object.quantity) : 0,
    };
  },

  toJSON(message: BeerReady): unknown {
    const obj: any = {};
    if (message.beer_id !== 0) {
      obj.beerId = Math.round(message.beer_id);
    }
    if (message.beer_name !== "") {
      obj.beerName = message.beer_name;
    }
    if (message.quantity !== 0) {
      obj.quantity = Math.round(message.quantity);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BeerReady>, I>>(base?: I): BeerReady {
    return BeerReady.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BeerReady>, I>>(object: I): BeerReady {
    const message = createBaseBeerReady();
    message.beer_id = object.beer_id ?? 0;
    message.beer_name = object.beer_name ?? "";
    message.quantity = object.quantity ?? 0;
    return message;
  },
};

function createBaseOrderComplete(): OrderComplete {
  return { order_id: "", beers: [], total_beers: 0 };
}

export const OrderComplete: MessageFns<OrderComplete> = {
  encode(message: OrderComplete, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.order_id !== "") {
      writer.uint32(10).string(message.order_id);
    }
    for (const v of message.beers) {
      BeerReady.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.total_beers !== 0) {
      writer.uint32(24).int32(message.total_beers);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OrderComplete {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrderComplete();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.order_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.beers.push(BeerReady.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.total_beers = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OrderComplete {
    return {
      order_id: isSet(object.orderId)
        ? globalThis.String(object.orderId)
        : isSet(object.order_id)
        ? globalThis.String(object.order_id)
        : "",
      beers: globalThis.Array.isArray(object?.beers) ? object.beers.map((e: any) => BeerReady.fromJSON(e)) : [],
      total_beers: isSet(object.totalBeers)
        ? globalThis.Number(object.totalBeers)
        : isSet(object.total_beers)
        ? globalThis.Number(object.total_beers)
        : 0,
    };
  },

  toJSON(message: OrderComplete): unknown {
    const obj: any = {};
    if (message.order_id !== "") {
      obj.orderId = message.order_id;
    }
    if (message.beers?.length) {
      obj.beers = message.beers.map((e) => BeerReady.toJSON(e));
    }
    if (message.total_beers !== 0) {
      obj.totalBeers = Math.round(message.total_beers);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OrderComplete>, I>>(base?: I): OrderComplete {
    return OrderComplete.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OrderComplete>, I>>(object: I): OrderComplete {
    const message = createBaseOrderComplete();
    message.order_id = object.order_id ?? "";
    message.beers = object.beers?.map((e) => BeerReady.fromPartial(e)) || [];
    message.total_beers = object.total_beers ?? 0;
    return message;
  },
};

function createBaseOrderEvent(): OrderEvent {
  return { beer_ready: undefined, order_complete: undefined };
}

export const OrderEvent: MessageFns<OrderEvent> = {
  encode(message: OrderEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.beer_ready !== undefined) {
      BeerReady.encode(message.beer_ready, writer.uint32(10).fork()).join();
    }
    if (message.order_complete !== undefined) {
      OrderComplete.encode(message.order_complete, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OrderEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrderEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.beer_ready = BeerReady.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.order_complete = OrderComplete.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OrderEvent {
    return {
      beer_ready: isSet(object.beerReady)
        ? BeerReady.fromJSON(object.beerReady)
        : isSet(object.beer_ready)
        ? BeerReady.fromJSON(object.beer_ready)
        : undefined,
      order_complete: isSet(object.orderComplete)
        ? OrderComplete.fromJSON(object.orderComplete)
        : isSet(object.order_complete)
        ? OrderComplete.fromJSON(object.order_complete)
        : undefined,
    };
  },

  toJSON(message: OrderEvent): unknown {
    const obj: any = {};
    if (message.beer_ready !== undefined) {
      obj.beerReady = BeerReady.toJSON(message.beer_ready);
    }
    if (message.order_complete !== undefined) {
      obj.orderComplete = OrderComplete.toJSON(message.order_complete);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OrderEvent>, I>>(base?: I): OrderEvent {
    return OrderEvent.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OrderEvent>, I>>(object: I): OrderEvent {
    const message = createBaseOrderEvent();
    message.beer_ready = (object.beer_ready !== undefined && object.beer_ready !== null)
      ? BeerReady.fromPartial(object.beer_ready)
      : undefined;
    message.order_complete = (object.order_complete !== undefined && object.order_complete !== null)
      ? OrderComplete.fromPartial(object.order_complete)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}